Added error bars to p-value interpolation
Performed another 1k x 51 simulations for the expected p-value and saved the full original set of data. Added standard deviation and standard error on the mean analysis on the graph to show uncertainties. The p-value is found to be 254 - arguably very close to our previous simulation 255. Improved plot aesthetics and scaling.